Написаны тесты на методы готового класса. Добавлено описание в README.
