+++ README.md
@@ -1,0 +1,8 @@
+Написаны юнит-тесты на методы класса из задания.
+В тестах используются параметризация и моки.
+В проекте использовано:
+ - maven 4.0
+ - java 11
+ - junit 4.13.2
+ - mockito 5.5.0
+ - jacoco 0.8.7
