+++ src/main/java/ru/yandex/praktikum/IngredientType.java
@@ -1,0 +1,11 @@
+package ru.yandex.praktikum;
+
+/**
+ * Перечисление с типами ингредиентов.
+ * SAUCE – соус
+ * FILLING – начинка
+ */
+public enum IngredientType {
+    SAUCE,
+    FILLING
+}
